test in separate file and seperate requirements
