+++ h160_ss58_bridge/requirements-test.txt
@@ -1,0 +1,3 @@
+-r requirements.txt
+pytest~=8.3.5
+eth-account~=0.13.5
